use global options instead of project-specific ones
