+++ Python/Product/Debugger/Debugger/IPythonDebugOptionsService.cs
@@ -65,5 +65,10 @@ namespace Microsoft.PythonTools.Debugger {
         /// True to use the legacy debugger. Default is false.
         /// </summary>
         bool UseLegacyDebugger { get; }
+
+        PresentationMode VariablePresentationForClasses { get; }
+        PresentationMode VariablePresentationForFunctions { get; }
+        PresentationMode VariablePresentationForProtected { get; }
+        PresentationMode VariablePresentationForSpecial { get; }
     }
 }
